requirements.txt
streamlit==1.25.0
python-dotenv==1.0.0
PyPDF2==3.0.1
streamlit-extras
langchain==0.0.208
openai==0.27.8
faiss-cpu==1.7.4
tiktoken
